[Insights] Add a frame stats mode to the timer panel. #rb Ionut.Matasaru #preflight 641311f8290c6e5d77d6dd06
[CL 24674993 by Catalin Dragoiu in ue5-main branch]

+++ Engine/Source/Developer/TraceInsights/Private/Insights/ViewModels/TimingGraphTrack.cpp
@@ -480,7 +480,7 @@ void FTimingGraphTrack::UpdateFrameStatsTimerSeries(FTimingGraphSeries& Series, 
 				return;
 			}
 
-			FramesProvider.EnumerateFrames(ETraceFrameType::TraceFrameType_Game, 0, FrameCount, [&Series](const TraceServices::FFrame& Frame)
+			FramesProvider.EnumerateFrames(ETraceFrameType::TraceFrameType_Game, (uint64) 0, (uint64) FrameCount, [&Series](const TraceServices::FFrame& Frame)
 				{
 					FTimingGraphSeries::FAtomicTimingEvent Event;
 					Event.FrameStartTime = Frame.StartTime;
